Explain the purpose in the REAdME
When I pointed some friends to this repo, without preliminatory explanation,
it wasn't completely clear why we developped it.
Hence, I try to explain it in our README, because I'm afraid other people
migh also easily miss the point.
(Said differently: it's clear for us because we know our needs, but it's not
 such a common situation)